requirements.txt
docling
hf-xet
transformers
google-genai
google-generativeai
python-dotenv
chromadb
pydantic-ai